pyproject.toml
[tool.poetry]
name = "rag-app"
version = "0.1.0"
description = "RAG App con extracción de PDFs, OCR e integración LangChain"
authors = ["Vicenç Delgado"]
readme = "README.md"
packages = [{ include = "src" }]  # Asegúrate de que la lógica de tu app está ahí o cambia esto

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
aiofiles = "^24.1.0"
aiohttp = "^3.12.14"
annotated-types = "^0.7.0"
attrs = "^25.3.0"
backoff = "^2.2.1"
beautifulsoup4 = "^4.13.4"
cffi = "^1.17.1"
chardet = "^5.2.0"
click = "^8.2.1"
cryptography = "^45.0.5"
dataclasses-json = "^0.6.7"
deprecated = "^1.2.18"
emoji = "^2.14.1"
et-xmlfile = "^2.0.0"
filetype = "^1.2.0"
fsspec = "^2025.5.1"
google-api-core = "^2.25.1"
google-auth = "^2.40.3"
google-cloud-vision = "^3.10.2"
grpcio = "^1.73.1"
grpcio-status = "^1.73.1"
httpx = "0.27.0"
huggingface-hub = "^0.33.4"
iopath = "^0.1.10"
jinja2 = "^3.1.6"
joblib = "^1.5.1"
jsonpatch = "^1.33"
langchain = "^0.1.20"  # ✅ Compatible con experimental 0.0.49
langchain-community = "^0.0.38"
langchain-core = "^0.1.53"
langchain-openai = "^0.1.7"
langchain-text-splitters = "^0.0.2"
langdetect = "^1.0.9"
langsmith = "^0.1.147"
langchain-experimental = "^0.0.49"
layoutparser = "^0.3.4"
lxml = "^6.0.0"
markdown = "^3.8.2"
marshmallow = "^3.26.1"
mpmath = "^1.3.0"
msg-parser = "^1.2.0"
multidict = "^6.6.3"
mypy-extensions = "^1.1.0"
nest-asyncio = "^1.6.0"
networkx = "^3.5"
nltk = "^3.9.1"
numpy = "^1.26.4"
omegaconf = "^2.3.0"
onnx = "^1.18.0"
onnxruntime = "^1.22.1"
opencv-python = "^4.11.0.86"
openai = "^1.95.1"
openpyxl = "^3.1.5"
orjson = "^3.10.18"
packaging = "^23.2"
pandas = "^2.3.1"
pdf2image = "^1.17.0"
pdfminer-six = "20221105"
pgvector = "^0.2.4"
pikepdf = "^9.9.0"
pillow = "^10.4.0"
pillow-heif = "^0.22.0"
portalocker = "^3.2.0"
proto-plus = "^1.26.1"
protobuf = "^6.31.1"
psycopg = { version = "^3.1.18", extras = ["binary"] }
pyasn1 = "^0.6.1"
pyasn1-modules = "^0.4.2"
pycocotools = "^2.0.10"
pycparser = "^2.22"
pydantic = "^2.11.7"
pydantic-core = "^2.33.2"
pyparsing = "^3.1.2"
pypandoc = "^1.15"
pypdf = "^5.8.0"
pypdfium2 = "^4.30.1"
pyreadline3 = "^3.5.4"
python-dateutil = "^2.9.0.post0"
python-docx = "^1.2.0"
python-dotenv = "^1.1.1"
python-iso639 = "^2025.2.18"
python-magic = "^0.4.27"
python-multipart = "^0.0.20"
python-pptx = "^0.6.23"
pytest = "^8.2"
pytesseract = "^0.3.13"
pytz = "^2025.2"
pywin32 = "^310"
pyyaml = "^6.0.2"
regex = "^2024.11.6"
rsa = "^4.9.1"
safetensors = "^0.5.3"
scipy = "^1.16.0"
six = "^1.17.0"
soupsieve = "^2.7"
sqlalchemy = "^2.0.41"
tabulate = "^0.9.0"
tenacity = "^8.5.0"
tiktoken = "^0.7.0"
timm = "^1.0.17"
tokenizers = "^0.21.2"
torch = "^2.7.1"
torchvision = "^0.22.1"
tqdm = "^4.67.1"
transformers = "^4.53.2"
typing-extensions = "^4.14.1"
typing-inspect = "^0.9.0"
typing-inspection = "^0.4.1"
tzdata = "^2025.2"
unstructured = "^0.13.7"
unstructured-client = "^0.38.1"
unstructured-inference = "^0.7.31"
unstructured-pytesseract = "^0.3.15"
urllib3 = "^2.5.0"
wrapt = "^1.17.2"
xlrd = "^2.0.2"
xlsxwriter = "^3.2.5"
yarl = "^1.20.1"
psycopg2-binary = "^2.9.10"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
langserve = "^0.0.36"
sse-starlette = "^3.0.2"
sentence-transformers = "^5.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
